+++ dwds/CHANGELOG.md
@@ -7,6 +7,7 @@
   for lazily created SDK configurations.
 - Fix an issue in reporting DevTools stats where the DevTools load time was
   not always recorded.
+- Add an `ide` query parameter to the Dart DevTools URL for analytics.
 
 **Breaking changes:**
 - `Dwds.start` and `ExpressionCompilerService` now take